Move RemoveOrphanSemantidbFiles to semanticdb-scalac core
